Add files via upload
This code is separate from the previous one. I got this from the tutorial link that Canvas supplied. I'm not sure if this is exactly the same as how prof presented it but it is in the right direction and I'm sure small adjustments will be made. This server program does involve multithreading, but doesn't spawn multiple clients at once as canvas says. When running the Server program, type "java Server 1234" then press enter. For the Client program, all you gotta do is run it then type "localhost" for hostname and "1234" for port. From there, I created a menu using switch statements that have the different functions. I've only completed 3/6 functions. For the client to leave the server, type "bye".

The main thing we need to figure out now is the spawning of the user-determined # of clients.